Add hacks to get around current limits on history updates.
At present theano histories can only be updated one time step at a time.
Certain histories (the discretized kernels) need to be updated fully computed
at the first time step, and then just read repeatedly. To achieve this, they
are forced to be computed in the initialization with `.set()` and then, and this
is the real hack, they are are accessed by indexing the underlying `._data`
directly, thus bypassing the logic that would try to compute them again.