[fsGIF] Fixed the super long 'compilation' times.
The problem was in choosing the length of the kernel: I was using .eval() at
every iteration, making it compile and throw away every time. Smaller time bins
meant more iterations, and therefore more time.
A secondary problem was when testing whether a NumPy array was a Theano object,
shim would recurse through the array, so a 2x2000 array would trigger 4000
checks. Shim was changed to terminate recursions when it hits NumPy arrays.