[fsGIF] Convert loaded parameters to arrays. Add load_parameters.
Array parameters are written to the parameter file as nested lists for
legibility. This commit ensures they are converted to NumPy arrays when loaded.

Add `load_parameters` for use of the code as an imported module. Code which
imports main.py now needs to add a call to this method explicitly specifying the
parameter file which would otherwise be given on the command line.