[mcmc] Add custom PyMC3 model allowing setup function
A problem with the MCMC runs was that my likelihood function needs to be
initialized before being evaluated (setting state variables + burnin), and as
this is not part of the computational graph it wasn't done. This commit adds a
custom PyMC3 model, which allows assigning a "setup" function to do this.

This commit also adds functionality to use the MLE as a starting point for the
MCMC. MLE is loaded from a file, which was written to by the new script
`calc_mle.py`.