[sgd] Remove duplication of parameters
posterior.params and gradient_descent.params both defined 'burnin' and 'datalen'
parameters (the former of which was called 'start' in gradient_descent.params).
This commit deprecates the values in gradient_descent.params
gradient_descent.params also defined a 'burnin' value, which was the burnin used
between batches. This was renamed 'batch_burnin' to prevent confusion.