[mle-activity] Add `update_params` function.
This function updates a base parameter set with values of another source 
dictionary. It differs from a plain update by calling `params_to_arrays` 
on the source, and reshaping the values to the same shape as the 
corresponding value in the base parameter set. This allows it to be used 
with the result from a parameter fit (FitCollection.result)